pyproject.toml
[project]
name = "voice-form-backend"
version = "0.1.0"
description = "FastAPI backend for voice form uploads"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.8",
  "uvicorn==0.34.0",
  "boto3==1.36.18",
  "python-dotenv==1.0.1",
  "python-multipart==0.0.20",
  "sqlalchemy==2.0.38",
  "psycopg2-binary==2.9.9",
]

[tool.uvicorn]
factory = false
